fix: resolve missing imports in copy commands
