+++ .omg/state/learn-watch.json
@@ -1,12 +1,12 @@
 {
   "last_session_id": "5ea30197-3847-496e-85cf-857924193301",
-  "last_event_key": "5ea30197-3847-496e-85cf-857924193301:transcript:c847adfbff0a41df372150d5c38a75e19331e2019be38b44c16553c07e08b1c6",
-  "last_prompted_session_id": "",
-  "last_reason": "short-session",
-  "last_prompted_at": "",
-  "last_user_message_count": 8,
-  "last_actionable_message_count": 1,
+  "last_event_key": "5ea30197-3847-496e-85cf-857924193301:transcript:6e0583841a8705f47eb0a6f147205e965fb5493c69b5049429b8882663844e3d",
+  "last_prompted_session_id": "5ea30197-3847-496e-85cf-857924193301",
+  "last_reason": "prompt-cooldown-active",
+  "last_prompted_at": "2026-04-22T09:19:09.509Z",
+  "last_user_message_count": 11,
+  "last_actionable_message_count": 2,
   "deep_interview_lock_active": false,
   "deep_interview_lock_source": "/home/dbado/sfDBTools/.omg/state/deep-interview.json",
-  "updated_at": "2026-04-22T07:48:37.208Z"
+  "updated_at": "2026-04-22T09:20:05.845Z"
 }
