+++ internal/app/copy/service.go
@@ -1,30 +1,12 @@
 package copy
 
 import (
-	"context"
-	"fmt"
-	"os"
-	"path/filepath"
-	"strings"
-	"time"
-
-	"sfdbtools/internal/app/backup/execution"
-	backuppath "sfdbtools/internal/app/backup/helpers/path"
-	"sfdbtools/internal/app/backup/model/types_backup"
-	copyexec "sfdbtools/internal/app/copy/execution"
-	profileconn "sfdbtools/internal/app/profile/connection"
 	"sfdbtools/internal/app/profile/helpers/loader"
-	"sfdbtools/internal/app/restore"
-	restoremodel "sfdbtools/internal/app/restore/model"
 	"sfdbtools/internal/domain"
 	appconfig "sfdbtools/internal/services/config"
 	applog "sfdbtools/internal/services/log"
-	"sfdbtools/internal/shared/compress"
 	"sfdbtools/internal/shared/consts"
-	"sfdbtools/internal/shared/database"
 	"sfdbtools/internal/shared/errorlog"
-	"sfdbtools/internal/ui/progress"
-	"sfdbtools/internal/ui/prompt"
 )
 
 // Service adalah orchestrator utama untuk fitur copy.
@@ -66,427 +48,4 @@ func (s *Service) LoadProfile(profileName, profileKey string, allowInteractive b
 		RequireProfile:   true,
 		AllowInteractive: allowInteractive,
 	})
-}
-
-// SelectDatabaseInteractive memunculkan picker untuk memilih database dari server.
-func (s *Service) SelectDatabaseInteractive(ctx context.Context, profile *domain.ProfileInfo) (string, error) {
-	client, err := profileconn.ConnectWithProfile(s.cfg, profile, "")
-	if err != nil {
-		return "", err
-	}
-	defer client.Close()
-
-	dbs, err := client.GetNonSystemDatabases(ctx)
-	if err != nil {
-		return "", err
-	}
-
-	if len(dbs) == 0 {
-		return "", fmt.Errorf("tidak ada database yang tersedia di server")
-	}
-
-	selected, _, err := prompt.SelectOne("Pilih database sumber:", dbs, 0)
-	return selected, err
-}
-
-// SelectTableInteractive memunculkan picker untuk memilih database dan tabel dari server.
-func (s *Service) SelectTableInteractive(ctx context.Context, profile *domain.ProfileInfo) (string, string, error) {
-	client, err := profileconn.ConnectWithProfile(s.cfg, profile, "")
-	if err != nil {
-		return "", "", err
-	}
-	defer client.Close()
-
-	// 1. Pilih DB
-	dbs, err := client.GetNonSystemDatabases(ctx)
-	if err != nil {
-		return "", "", err
-	}
-	dbName, _, err := prompt.SelectOne("Pilih database:", dbs, 0)
-	if err != nil {
-		return "", "", err
-	}
-
-	// 2. Ambil list tabel
-	query := fmt.Sprintf("SHOW TABLES FROM `%s` ", dbName)
-	rows, err := client.QueryContextWithRetry(ctx, query)
-	if err != nil {
-		return "", "", err
-	}
-	defer rows.Close()
-
-	var tables []string
-	for rows.Next() {
-		var t string
-		if err := rows.Scan(&t); err == nil {
-			tables = append(tables, t)
-		}
-	}
-
-	if len(tables) == 0 {
-		return "", "", fmt.Errorf("tidak ada tabel di database '%s'", dbName)
-	}
-
-	// 3. Pilih Tabel
-	tableName, _, err := prompt.SelectOne(fmt.Sprintf("Pilih tabel di %s:", dbName), tables, 0)
-	return dbName, tableName, err
-}
-
-// CopyDatabase melakukan penyalinan database utuh.
-func (s *Service) CopyDatabase(ctx context.Context, profile *domain.ProfileInfo, sourceDB, targetDB string, schemaOnly, useDisk, force, backupFirst, nonInteractive bool) (string, error) {
-	// 1. Connect to DB
-	client, err := profileconn.ConnectWithProfile(s.cfg, profile, "")
-	if err != nil {
-		return "", fmt.Errorf("gagal koneksi ke database: %w", err)
-	}
-	defer client.Close()
-
-	// 2. Pre-flight checks
-	exists, err := client.CheckDatabaseExists(ctx, sourceDB)
-	if err != nil {
-		return "", err
-	}
-	if !exists {
-		return "", fmt.Errorf("database sumber '%s' tidak ditemukan", sourceDB)
-	}
-
-	if targetDB == "" {
-		if nonInteractive {
-			targetDB = fmt.Sprintf("%s_copy_%s", sourceDB, time.Now().Format("20060102"))
-			s.log.Infof("Nama database target otomatis: %s", targetDB)
-		} else {
-			targetDB, err = prompt.AskText("Masukkan nama database target:", prompt.WithDefault(fmt.Sprintf("%s_copy_%s", sourceDB, time.Now().Format("20060102"))))
-			if err != nil {
-				return "", err
-			}
-		}
-	}
-
-	if strings.EqualFold(sourceDB, targetDB) {
-		return "", fmt.Errorf("database target tidak boleh sama dengan database sumber")
-	}
-
-	targetExists, err := client.CheckDatabaseExists(ctx, targetDB)
-	if err != nil {
-		return "", err
-	}
-
-	if targetExists {
-		if !nonInteractive {
-			// Mode Interaktif: Konfirmasi berlapis
-			s.log.Warnf("PERINGATAN: Database target '%s' sudah ada!", targetDB)
-
-			// Jika belum ada flag yang menspesifikasi tindakan, tanya user
-			if !force && !backupFirst {
-				choice, _, err := prompt.SelectOne("Database target sudah ada. Apa yang ingin Anda lakukan?",
-					[]string{"Batalkan", "Backup dulu baru timpa (Sangat Disarankan)", "Timpa langsung (Data lama hilang!)"}, 0)
-				if err != nil || choice == "Batalkan" {
-					return "", fmt.Errorf("operasi dibatalkan")
-				}
-				if choice == "Backup dulu baru timpa (Sangat Disarankan)" {
-					backupFirst = true
-				} else {
-					force = true
-				}
-			}
-
-			// Konfirmasi Akhir
-			if backupFirst {
-				confirm, err := prompt.Confirm(fmt.Sprintf("Yakin ingin membackup lalu menimpa database '%s'?", targetDB), true)
-				if err != nil || !confirm {
-					return "", fmt.Errorf("operasi dibatalkan")
-				}
-			} else if force {
-				// Konfirmasi Berlapis untuk Overwrite tanpa Backup
-				s.log.Warnf("!!! PERHATIAN !!!")
-				s.log.Warnf("Anda memilih untuk MENIMPA database '%s' TANPA backup.", targetDB)
-				confirm1, _ := prompt.Confirm("Apakah Anda sadar data lama di target akan hilang permanen?", false)
-				if !confirm1 {
-					return "", fmt.Errorf("operasi dibatalkan")
-				}
-				confirm2, _ := prompt.Confirm("Konfirmasi terakhir: Benar-benar ingin menimpa tanpa backup?", false)
-				if !confirm2 {
-					return "", fmt.Errorf("operasi dibatalkan")
-				}
-			}
-		} else {
-			// Mode Non-Interaktif: Cek flag
-			if !force && !backupFirst {
-				return "", fmt.Errorf("database target '%s' sudah ada. Gunakan --force atau --backup-first", targetDB)
-			}
-		}
-
-		// Jalankan backup target jika diminta
-		if backupFirst {
-			s.log.Infof("Menjalankan backup pengamanan untuk target: %s", targetDB)
-			if err := s.runSafetyBackup(ctx, profile, client, targetDB); err != nil {
-				return "", fmt.Errorf("gagal melakukan backup pengamanan: %w", err)
-			}
-		}
-	}
-
-	// 3. Determine method
-	if !useDisk && !nonInteractive {
-		choice, _, err := prompt.SelectOne("Pilih metode penyalinan:", []string{"Direct Stream (Cepat, RAM-based)", "Disk-based (Aman, Dump file)"}, 0)
-		if err == nil && choice == "Disk-based (Aman, Dump file)" {
-			useDisk = true
-		}
-	}
-
-	methodName := "Piping"
-	if useDisk {
-		methodName = "Disk-based"
-	}
-	s.log.Infof("Memulai copy database: %s -> %s [Metode: %s]", sourceDB, targetDB, methodName)
-
-	// 4. Create target DB if not exists
-	if err := client.CreateDatabaseIfNotExists(ctx, targetDB); err != nil {
-		return "", err
-	}
-
-	// 5. Execution
-	if useDisk {
-		if err := s.executeDiskCopy(ctx, profile, client, sourceDB, targetDB, schemaOnly); err != nil {
-			return "", err
-		}
-	} else {
-		if err := copyexec.ExecutePiping(ctx, s.log, copyexec.PipingOptions{
-			Profile:      profile,
-			SourceDB:     sourceDB,
-			TargetDB:     targetDB,
-			SchemaOnly:   schemaOnly,
-			BaseDumpArgs: s.cfg.Backup.MysqlDumpArgs,
-		}); err != nil {
-			return "", err
-		}
-	}
-
-	return targetDB, nil
-}
-
-// runSafetyBackup melakukan backup cepat ke direktori default config sebelum ditimpa.
-func (s *Service) runSafetyBackup(ctx context.Context, profile *domain.ProfileInfo, client *database.Client, dbName string) error {
-	opts := &types_backup.BackupDBOptions{
-		Profile: *profile,
-		Mode:    consts.ModeSingle,
-		DBName:  dbName,
-		Ticket:  "SAFETY_AUTO_BACKUP",
-	}
-
-	// Gunakan config default untuk kompresi agar cepat
-	opts.Compression.Enabled = true
-	opts.Compression.Type = consts.CompressionTypeGzip
-
-	// Generate path
-	hostname := profile.DBInfo.Host
-	filename, _ := backuppath.GenerateBackupFilename(dbName, consts.ModeSingle, hostname, compress.CompressionType(consts.CompressionTypeGzip), false, false)
-
-	outputDir := s.cfg.Backup.Output.BaseDirectory
-	if outputDir == "" {
-		outputDir = filepath.Join(os.TempDir(), "sfdbtools_safety")
-		_ = os.MkdirAll(outputDir, 0755)
-	}
-	outputPath := filepath.Join(outputDir, filename)
-
-	s.log.Infof("Backup disimpan di: %s", outputPath)
-
-	eng := execution.New(s.log, s.cfg, opts, s.errLog).WithDependencies(client, nil, nil, nil, nil)
-	_, err := eng.ExecuteAndBuildBackup(ctx, types_backup.BackupExecutionConfig{
-		DBName:       dbName,
-		OutputPath:   outputPath,
-		IsMultiDB:    false,
-		TotalDBFound: 1,
-	})
-	return err
-}
-
-func (s *Service) executeDiskCopy(ctx context.Context, profile *domain.ProfileInfo, client *database.Client, sourceDB, targetDB string, schemaOnly bool) error {
-	workdir := filepath.Join(os.TempDir(), fmt.Sprintf("sfdbtools_copy_%d", time.Now().Unix()))
-	if err := os.MkdirAll(workdir, 0755); err != nil {
-		return err
-	}
-	defer os.RemoveAll(workdir)
-
-	// A. Backup
-	opts := &types_backup.BackupDBOptions{
-		Profile: *profile,
-		Mode:    consts.ModeSingle,
-		DBName:  sourceDB,
-	}
-	opts.Filter.ExcludeData = schemaOnly
-
-	filename, err := backuppath.GenerateBackupFilename(sourceDB, consts.ModeSingle, profile.DBInfo.Host, compress.CompressionType(consts.CompressionTypeGzip), false, false)
-	if err != nil {
-		return err
-	}
-	outputPath := filepath.Join(workdir, filename)
-
-	eng := execution.New(s.log, s.cfg, opts, s.errLog).WithDependencies(client, nil, nil, nil, nil)
-	_, err = eng.ExecuteAndBuildBackup(ctx, types_backup.BackupExecutionConfig{
-		DBName:       sourceDB,
-		OutputPath:   outputPath,
-		IsMultiDB:    false,
-		TotalDBFound: 1,
-	})
-	if err != nil {
-		return fmt.Errorf("gagal backup source: %w", err)
-	}
-
-	// B. Restore
-	restOpts := &restoremodel.RestoreSingleOptions{
-		Profile:     *profile,
-		File:        outputPath,
-		TargetDB:    targetDB,
-		Force:       true,
-		StopOnError: true,
-		SkipBackup:  true, // Karena ini copy, kita asumsikan target bisa ditimpa
-		DropTarget:  true,
-	}
-
-	restSvc := restore.NewRestoreService(s.log, s.cfg, restOpts)
-	if err := restSvc.SetupRestoreSession(ctx); err != nil {
-		return fmt.Errorf("gagal setup restore: %w", err)
-	}
-	defer restSvc.Close()
-
-	_, err = restSvc.ExecuteRestoreSingle(ctx)
-	return err
-}
-
-// CopyTable melakukan penyalinan tabel spesifik.
-func (s *Service) CopyTable(ctx context.Context, profile *domain.ProfileInfo, sourceDB, sourceTable, targetDB, targetTable string, schemaOnly, force, backupFirst, nonInteractive bool) (string, string, error) {
-	client, err := profileconn.ConnectWithProfile(s.cfg, profile, "")
-	if err != nil {
-		return "", "", fmt.Errorf("gagal koneksi ke database: %w", err)
-	}
-	defer client.Close()
-
-	// Validation
-	exists, err := s.validateTableExists(ctx, client, sourceDB, sourceTable)
-	if err != nil || !exists {
-		return "", "", fmt.Errorf("tabel sumber %s.%s tidak ditemukan", sourceDB, sourceTable)
-	}
-
-	if targetDB == "" {
-		targetDB = sourceDB
-	}
-
-	targetExists, err := s.validateTableExists(ctx, client, targetDB, targetTable)
-	if err != nil {
-		return "", "", err
-	}
-
-	if targetExists {
-		if !nonInteractive {
-			s.log.Warnf("PERINGATAN: Tabel target '%s.%s' sudah ada!", targetDB, targetTable)
-
-			if !force && !backupFirst {
-				choice, _, err := prompt.SelectOne("Tabel target sudah ada. Apa yang ingin Anda lakukan?",
-					[]string{"Batalkan", "Backup dulu baru timpa", "Timpa langsung (Berisiko!)"}, 0)
-				if err != nil || choice == "Batalkan" {
-					return "", "", fmt.Errorf("operasi dibatalkan")
-				}
-				if choice == "Backup dulu baru timpa" {
-					backupFirst = true
-				} else {
-					force = true
-				}
-			}
-
-			if backupFirst {
-				confirm, err := prompt.Confirm(fmt.Sprintf("Yakin ingin membackup lalu menimpa tabel '%s.%s'?", targetDB, targetTable), true)
-				if err != nil || !confirm {
-					return "", "", fmt.Errorf("operasi dibatalkan")
-				}
-			} else if force {
-				s.log.Warnf("!!! PERHATIAN !!!")
-				confirm1, _ := prompt.Confirm(fmt.Sprintf("Yakin ingin MENIMPA tabel '%s.%s' TANPA backup?", targetDB, targetTable), false)
-				if !confirm1 {
-					return "", "", fmt.Errorf("operasi dibatalkan")
-				}
-				confirm2, _ := prompt.Confirm("Benar-benar yakin? Data lama akan hilang.", false)
-				if !confirm2 {
-					return "", "", fmt.Errorf("operasi dibatalkan")
-				}
-			}
-		} else {
-			if !force && !backupFirst {
-				return "", "", fmt.Errorf("tabel target '%s.%s' sudah ada. Gunakan --force atau --backup-first", targetDB, targetTable)
-			}
-		}
-
-		if backupFirst {
-			s.log.Infof("Menjalankan backup pengamanan untuk tabel: %s.%s", targetDB, targetTable)
-			if err := s.runSafetyTableBackup(ctx, profile, client, targetDB, targetTable); err != nil {
-				return "", "", fmt.Errorf("gagal melakukan backup pengamanan tabel: %w", err)
-			}
-		}
-	}
-
-	s.log.Infof("Memulai copy tabel: %s.%s -> %s.%s", sourceDB, sourceTable, targetDB, targetTable)
-
-	spin := progress.NewSpinnerWithElapsed(fmt.Sprintf("Copying table %s.%s", sourceDB, sourceTable))
-	spin.Start()
-	defer spin.Stop()
-
-	if force || backupFirst {
-		dropSQL := fmt.Sprintf("DROP TABLE IF EXISTS `%s`.`%s` ", targetDB, targetTable)
-		_, _ = client.ExecContextWithRetry(ctx, dropSQL)
-	}
-
-	createSQL := fmt.Sprintf("CREATE TABLE IF NOT EXISTS `%s`.`%s` LIKE `%s`.`%s` ", targetDB, targetTable, sourceDB, sourceTable)
-	if _, err := client.ExecContextWithRetry(ctx, createSQL); err != nil {
-		return "", "", fmt.Errorf("gagal membuat struktur tabel target: %w", err)
-	}
-
-	if !schemaOnly {
-		insertSQL := fmt.Sprintf("INSERT INTO `%s`.`%s` SELECT * FROM `%s`.`%s` ", targetDB, targetTable, sourceDB, sourceTable)
-		if _, err := client.ExecContextWithRetry(ctx, insertSQL); err != nil {
-			return "", "", fmt.Errorf("gagal menyalin data tabel: %w", err)
-		}
-	}
-
-	return targetDB, targetTable, nil
-}
-
-func (s *Service) runSafetyTableBackup(ctx context.Context, profile *domain.ProfileInfo, client *database.Client, dbName, tableName string) error {
-	opts := &types_backup.BackupDBOptions{
-		Profile: *profile,
-		Mode:    consts.ModeSingle,
-		DBName:  dbName,
-		Ticket:  "SAFETY_AUTO_BACKUP_TABLE",
-	}
-
-	hostname := profile.DBInfo.Host
-	filename, _ := backuppath.GenerateBackupFilename(dbName+"_"+tableName, consts.ModeSingle, hostname, compress.CompressionType(consts.CompressionTypeGzip), false, false)
-
-	outputDir := s.cfg.Backup.Output.BaseDirectory
-	if outputDir == "" {
-		outputDir = filepath.Join(os.TempDir(), "sfdbtools_safety")
-		_ = os.MkdirAll(outputDir, 0755)
-	}
-	outputPath := filepath.Join(outputDir, filename)
-
-	s.log.Infof("Backup tabel disimpan di: %s", outputPath)
-
-	eng := execution.New(s.log, s.cfg, opts, s.errLog).WithDependencies(client, nil, nil, nil, nil)
-
-	// Gunakan argumen mysqldump untuk tabel spesifik
-	_, err := eng.ExecuteAndBuildBackup(ctx, types_backup.BackupExecutionConfig{
-		DBName:       dbName,
-		OutputPath:   outputPath,
-		IsMultiDB:    false,
-		TotalDBFound: 1,
-	})
-	return err
-}
-
-func (s *Service) validateTableExists(ctx context.Context, client *database.Client, dbName, tableName string) (bool, error) {
-	query := "SELECT TABLE_NAME FROM information_schema.TABLES WHERE TABLE_SCHEMA = ? AND TABLE_NAME = ? LIMIT 1"
-	rows, err := client.QueryContextWithRetry(ctx, query, dbName, tableName)
-	if err != nil {
-		return false, err
-	}
-	defer rows.Close()
-	return rows.Next(), nil
 }
